refactor(web): split v4 transition service by data source
Move SDK/events_core and system.query_log logic into sibling modules and
keep the tRPC router as a thin adapter. No behavior change.
